更新坦克数值对齐文档规范，实现 PVP 双人对战、比赛设置、结算对比等缺失功能
- 修正 8 种坦克转向速度为 坦克对象.txt 规定值 (20-60)
- fireRate/bulletDuration 改为 double 类型，支持 0.5s/5.5s 精度
- 新增 PVP 双人对战模式（P1/P2 独立选坦克、模式选择对话框）
- 新增 PVP 比赛设置面板（比赛时间、加时赛、超时惩罚、7种作弊开关）
- 新增战斗结算数据对比（开火数/造成伤害/受到伤害 P1 vs P2/AI）
- 新增改装碎片互换（2特定碎片→1通用碎片）
- 修正抽卡新坦克概率分布（初始两坦克各 20%，其余 60% 随机）
- 修正子弹命中时 damageDealt 追踪

+++ src/model/manager/ElementFactory.java
@@ -53,7 +53,7 @@ public class ElementFactory {
         Bullet b = new Bullet(x, y, (int)direction, owner);
         b.setDamage(owner.getAttack());
         b.setSpeed((int)(owner.getBulletSpeed() * TankData.BULLET_SPEED_FACTOR));
-        b.setLifeTime(owner.getBulletDuration() * 1000);
+        b.setLifeTime((int)(owner.getBulletDuration() * 1000));
         b.setRebound(true);
         b.setMaxRebounds(0);
         b.setLaser(false);
@@ -70,7 +70,7 @@ public class ElementFactory {
             Bullet b = new Bullet(x, y, (int)((bulletDir % 360 + 360) % 360), owner);
             b.setDamage((int)(owner.getAttack() * damageMult));
             b.setSpeed((int)(owner.getBulletSpeed() * TankData.BULLET_SPEED_FACTOR));
-            b.setLifeTime(owner.getBulletDuration() * 1000);
+            b.setLifeTime((int)(owner.getBulletDuration() * 1000));
             b.setRebound(false);
             b.setMaxRebounds(0);
             b.setLaser(false);
